+++ src/de/hhu/bsinfo/dxcompute/ms/MasterSlaveConfigurationValues.java
@@ -14,7 +14,7 @@ public class MasterSlaveConfigurationValues {
 	 */
 	public static class Service {
 		public static final Pair<String, String> ROLE = new Pair<String, String>("Role", "None");
-		public static final Pair<String, Integer> COMPUTE_GROUP_ID = new Pair<String, Integer>("ComputeGroupId", 0);
+		public static final Pair<String, Short> COMPUTE_GROUP_ID = new Pair<String, Short>("ComputeGroupId", (short) 0);
 		public static final Pair<String, Integer> PING_INTERVAL_MS =
 				new Pair<String, Integer>("PingIntervalMs", 1000);
